Update Serverbotv2.py
Added a list players command and updated the subprocess_handle config to allow for a stdout so we can fetch info outputted to the terminal.

This however will mess with what you see in the terminal as it's all being sent to stdout. So I added some more print statements to replace that.